feat: [#3] Add EditCard Logic
