+++ WorkUp/WorkUp/View/ListView/ListView.swift
@@ -30,7 +30,11 @@ struct ListView: View {
             }
             LazyVGrid(columns: columns, spacing: 25) {
                 ForEach(Array(cards.enumerated()), id: \.element) { index, item in
-                    ListCardCell(card: item, cardIndex: index)
+                    NavigationLink {
+                        NewCardView(card: item, question: item.question, answer: item.answer, isEdit: true)
+                    } label: {
+                        ListCardCell(card: item, cardIndex: index)
+                    }
                 }
             }
             .padding(.horizontal)
@@ -53,57 +57,6 @@ struct ListView: View {
                 .frame(width: 30, height: 30)
         }))
         
-    }
-}
-
-struct ListCardCell: View {
-    @Environment(\.modelContext) var modelContext
-    @State private var showAlert = false
-
-    let card: NewCard
-    let cardIndex: Int
-    
-    var body: some View {
-        VStack(alignment:.leading, spacing: 0) {
-            HStack(spacing: 0) {
-                Text("\(String(format: "%02d", cardIndex + 1))")
-                    .font(.system(size: 25, weight: .bold))
-                    .padding(.leading, 14)
-                Spacer()
-                Button {
-                    showAlert = true
-                } label: {
-                    Image(systemName: "trash")
-                        .resizable()
-                        .scaledToFit()
-                        .foregroundStyle(.white)
-                        .frame(width: 20, height: 24)
-                        .padding(.trailing, 14)
-                }
-                .alert(isPresented: $showAlert) {
-                    Alert(
-                        title: Text("카드를 삭제할까요?"),
-                        message: Text("삭제된 카드는 복구되지 않습니다."),
-                        primaryButton: .destructive(Text("삭제")) {
-                            modelContext.delete(card)
-                        },
-                        secondaryButton: .cancel(Text("취소"))
-                    )
-                }
-            }
-            .frame(height: 49)
-            .background(Color(hex: "53E7FB"))
-            Text("\(card.question)")
-                .font(.system(size: 20, weight: .bold))
-                .lineSpacing(14)
-                .foregroundStyle(.white)
-                .padding(.horizontal,14)
-                .padding(.top, 11)
-            Spacer()
-        }
-        .frame(width: 161, height: 217)
-        .background(Color(hex: "232323"))
-        .clipShape(RoundedRectangle(cornerRadius: 13))
     }
 }
